Created helper file for functions
